feat: upload files

+++ server/src/routes/memories.ts
@@ -2,8 +2,15 @@ import { FastifyInstance } from 'fastify'
 import { prisma } from '../db/prismaClient'
 import z from 'zod'
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -18,7 +25,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -31,7 +38,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     })
 
-    return memory
+    // eslint-disable-next-line eqeqeq
+    if (!memory.isPublic && memory.id != request.user.sub) {
+      return reply.status(401).send()
+    }
   })
 
   app.post('/memories', async (request) => {
@@ -57,7 +67,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -72,9 +82,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    console.log(request.body)
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
 
-    const memory = prisma.memory.update({
+    // eslint-disable-next-line eqeqeq
+    if (memory.id != request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -82,21 +101,31 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '89b6bd70-247a-4a8e-8a64-786a4f42a6ab',
+        userId: request.user.sub,
       },
     })
 
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
 
-    const memory = await prisma.memory.delete({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.id != request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.delete({
       where: {
         id,
       },
